migrate the test suite to integration tests
